# server_backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: SESA News & Updates API
  version: 1.0.0
  description: API for managing news and updates for the SESA website

servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.sesa-web.com/api
    description: Production server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    csrfToken:
      type: apiKey
      in: header
      name: CSRF-Token

  schemas:
    NewsUpdate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          maxLength: 255
        content:
          type: string
        type:
          type: string
          enum: [neuigkeiten, releases, frameworks, ankündigungen]
        version:
          type: string
          maxLength: 50
          nullable: true
        published_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - title
        - content
        - type

    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 6
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string

paths:
  /auth/login:
    post:
      summary: Login to get access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /news:
    get:
      summary: Get all news updates
      tags:
        - News
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [news, release, framework, announcement]
          description: Filter news by type
      responses:
        '200':
          description: List of news updates
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NewsUpdate'

    post:
      summary: Create a new news update
      tags:
        - News
      security:
        - BearerAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsUpdate'
      responses:
        '201':
          description: News update created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/NewsUpdate'
        '401':
          description: Unauthorized
        '403':
          description: Invalid CSRF token

  /news/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        description: News update ID
    
    get:
      summary: Get a specific news update
      tags:
        - News
      responses:
        '200':
          description: News update found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/NewsUpdate'
        '404':
          description: News update not found

    put:
      summary: Update a news update
      tags:
        - News
      security:
        - BearerAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsUpdate'
      responses:
        '200':
          description: News update modified
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/NewsUpdate'
        '401':
          description: Unauthorized
        '403':
          description: Invalid CSRF token
        '404':
          description: News update not found

    delete:
      summary: Delete a news update
      tags:
        - News
      security:
        - BearerAuth: []
        - csrfToken: []
      responses:
        '204':
          description: News update deleted
        '401':
          description: Unauthorized
        '403':
          description: Invalid CSRF token
        '404':
          description: News update not found 